+++ tests/_craft/config/general.php
@@ -1,8 +1,11 @@
 <?php
 
+use craft\helpers\App;
+
 return [
 	'devMode' => true,
 	'postLoginRedirect' => '/',
 	'requireUserAgentAndIpForSession' => false,
-	'securityKey' => getenv('SECURITY_KEY'),
+	'securityKey' => App::env('SECURITY_KEY'),
+	'enableCsrfProtection' => false,
 ];