+++ tests/_data/config/multiple-link.php
@@ -1,0 +1,6 @@
+<?php
+
+return [
+    'authenticationEmailSubject' => 'Here is your magic login link',
+    'linkAccessLimit' => 3,
+];