+++ tests/_data/config/single-link.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+	'authenticationEmailSubject' => 'Here is your magic login link',
+];