+++ tests/_data/magiclogin_userrecord.php
@@ -2,15 +2,24 @@
 
 return [
 	'pending_user' => [
-		'id' => 2,
+		// 'id' => 1,
 		'username' => 'test',
 		'email' => 'test@example.com',
-		'pending' => 1,
+		'active' => false,
+		'uid' => '929e0997-03fa-401e-b745-99480e53bf97',
 	],
 	'creode_test_user' => [
-		'id' => 2,
+		// 'id' => 2,
 		'username' => 'test',
 		'email' => 'creode-test@example.com',
-		'pending' => 1,
+		'active' => false,
+		'uid' => 'fe432693-99c9-44f6-9902-f09e246e2ca8',
+	],
+	'magic_link_expiry_user' => [
+		// 'id' => 10,
+		'username' => 'magic-login-expiry',
+		'email' => 'magic-login-expiry@example.com',
+		'uid' => '16b52ee5-4386-4fed-991e-63c4a06a7fd5',
+		'active' => false,
 	],
 ];